requirements.txt
Django==5.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
celery==5.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
python-telegram-bot==20.8
